+++ libs/search-widget/src/widgets/chat-widget/ChatWidget.svelte
@@ -22,6 +22,7 @@
   export let client = 'widget';
   export let state: KBStates = 'PUBLISHED';
   export let standalone = false;
+  export let proxy = false;
   export let cssPath = '';
   export let prompt = '';
 
@@ -62,6 +63,7 @@
         kbSlug: kbslug,
         account,
         standalone,
+        proxy,
       },
       state,
       { prompt },